Server: Use C++ standard types

+++ Server/ItemManager/SharedMem.cpp
@@ -41,12 +41,12 @@ CSharedMemQueue::~CSharedMemQueue()
 		CloseHandle(m_hMMFile);
 }
 
-bool CSharedMemQueue::InitializeMMF(DWORD dwOffsetsize, int maxcount, LPCTSTR lpname, bool bCreate)
+bool CSharedMemQueue::InitializeMMF(uint32_t dwOffsetsize, int maxcount, LPCTSTR lpname, bool bCreate)
 {
 	if (maxcount < 1)
 		return false;
 
-	DWORD dwfullsize = dwOffsetsize * maxcount + sizeof(_SMQ_HEADER);
+	uint32_t dwfullsize = dwOffsetsize * maxcount + sizeof(_SMQ_HEADER);
 
 	m_nMaxCount = maxcount;
 	m_wOffset = dwOffsetsize;
@@ -72,7 +72,7 @@ bool CSharedMemQueue::InitializeMMF(DWORD dwOffsetsize, int maxcount, LPCTSTR lp
 
 	m_bMMFCreate = bCreate;
 	m_pHeader = (_SMQ_HEADER*) m_lpMMFile;
-	m_lReference = (LONG) (m_lpMMFile + sizeof(_SMQ_HEADER));		// 초기 위치 셋팅
+	m_lReference = (uintptr_t) (m_lpMMFile + sizeof(_SMQ_HEADER));		// 초기 위치 셋팅
 
 	if (bCreate)
 	{
@@ -88,7 +88,7 @@ bool CSharedMemQueue::InitializeMMF(DWORD dwOffsetsize, int maxcount, LPCTSTR lp
 
 int CSharedMemQueue::PutData(char* pBuf, int size)
 {
-	BYTE BlockMode;
+	uint8_t BlockMode;
 	int index = 0, temp_rear = 0;
 
 	if (size > static_cast<int>(m_wOffset))
@@ -113,7 +113,7 @@ int CSharedMemQueue::PutData(char* pBuf, int size)
 		return SMQ_WRITING;
 	}
 
-	LONG pQueue = m_lReference + (m_pHeader->Rear * m_wOffset);
+	uintptr_t pQueue = m_lReference + (m_pHeader->Rear * m_wOffset);
 
 	BlockMode = GetByte((char*) pQueue, index);
 	if (BlockMode == WR
@@ -143,7 +143,7 @@ int CSharedMemQueue::PutData(char* pBuf, int size)
 int CSharedMemQueue::GetData(char* pBuf)
 {
 	int index = 0, size = 0, temp_front = 0;
-	BYTE BlockMode;
+	uint8_t BlockMode;
 
 	if (m_pHeader->FrontMode == R)
 		return SMQ_READING;
@@ -162,7 +162,7 @@ int CSharedMemQueue::GetData(char* pBuf)
 		return SMQ_READING;
 	}
 
-	LONG pQueue = m_lReference + (m_pHeader->Front * m_wOffset);
+	uintptr_t pQueue = m_lReference + (m_pHeader->Front * m_wOffset);
 
 	index = 0;
 	BlockMode = GetByte((char*) pQueue, index);
